Consolidate capability mixins into one module with a single Capability base
Problem: Capability mixins were spread across three modules with two
independent Capability base classes: sources/capabilities.py and
powermeters/capabilities.py each defined their own Capability(ABC), while
the DAQ mixins in daq/daqdevice.py inherited bare ABC with no shared base
at all. Capability introspection (capabilities() / hasCapability()) was
duplicated verbatim on LaserSourceDevice and PowerMeterDevice, each bound
to its family's local Capability, and DAQ devices had no introspection.

Solution: Move every capability mixin into a single top-level module,
hardwarelibrary/capabilities.py, all subclassing one Capability base. The
DAQ enums (InputSource, TriggerSource, SampleClock, DAQNotification) move
with their mixins. Hoist capabilities() / hasCapability() onto
PhysicalDevice, bound to the single Capability, so every device -- DAQ
included -- gets capability introspection, and the duplicated methods are
removed (LaserSourceDevice becomes a pure marker; PowerMeterDevice keeps
only its power hooks).

The per-family capabilities.py / daqdevice.py files are deleted. All
internal imports point at hardwarelibrary.capabilities; the family package
__init__s still re-export their own mixins, so existing consumers such as
`from hardwarelibrary.daq import AnalogIOCapability` keep working.

Docs updated: capabilities are documented once on the Core page; the
per-family automodule sections and their file-path references are removed.

Full suite: 470 passed, 247 skipped, no regressions.
