+++ hardwarelibrary/powermeters/__init__.py
@@ -1,5 +1,5 @@
 
-from .capabilities import Capability, WavelengthCalibrationCapability, AutoScaleCapability, ScaleCapability
+from hardwarelibrary.capabilities import Capability, WavelengthCalibrationCapability, AutoScaleCapability, ScaleCapability
 from .powermeterdevice import PowerMeterDevice
 from .integradevice import IntegraDevice
 from .fieldmasterdevice import FieldMasterDevice, DebugFieldMasterDevice
